% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bike_cities}
\alias{bike_cities}
\title{List of cities currently included in bikedata}
\usage{
bike_cities()
}
\value{
A \code{data.frame} of cities, abbreviations, and names of bike
systems currently able to be accessed.
}
\description{
List of cities currently included in bikedata
}
\examples{
bike_cities ()
}
