% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-stats.R
\name{bike_daily_trips}
\alias{bike_daily_trips}
\title{Extract daily trip counts for all stations}
\usage{
bike_daily_trips(bikedb)
}
\arguments{
\item{bikedb}{Path to the SQLite3 database}
}
\value{
A \code{data.frame} containing daily dates and total numbers of trips
}
\description{
Extract daily trip counts for all stations
}
