% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-stats.R
\name{bike_datelimits}
\alias{bike_datelimits}
\title{Extract date-time limits from trip database}
\usage{
bike_datelimits(bikedb, city)
}
\arguments{
\item{bikedb}{A string containing the path to the SQLite3 database.
If no directory specified, it is presumed to be in \code{tempdir()}.}

\item{city}{If given, date limits are calculated only for trips in 
that city.}
}
\value{
A vector of 2 elements giving the date-time of the first and last
trips
}
\description{
Extract date-time limits from trip database
}
