% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-stats.R
\name{bike_demographic_data}
\alias{bike_demographic_data}
\title{Static summary of which systems provide demographic data}
\usage{
bike_demographic_data()
}
\value{
A \code{data.frame} detailing the kinds of demographc data provided
by the different systems
}
\description{
Static summary of which systems provide demographic data
}
\examples{
bike_demographic_data ()
}
