% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-stats.R
\name{bike_latest_files}
\alias{bike_latest_files}
\title{Check whether files in database are the latest published files}
\usage{
bike_latest_files(bikedb)
}
\arguments{
\item{bikedb}{A string containing the path to the SQLite3 database.
If no directory specified, it is presumed to be in \code{tempdir()}.}
}
\value{
A named vector of binary values: TRUE is files in \code{bikedb} are
the latest versions; otherwise FALSE, in which case \code{store_bikedata}
could be run to update the database.
}
\description{
Check whether files in database are the latest published files
}
