% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-test-data.R
\name{bike_rm_test_data}
\alias{bike_rm_test_data}
\title{Removes test data written with 'bike_write_test_data()'}
\usage{
bike_rm_test_data(data_dir = tempdir())
}
\arguments{
\item{data_dir}{Directory in which data were extracted.}
}
\description{
Removes test data written with 'bike_write_test_data()'
}
