% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{bike_stations}
\alias{bike_stations}
\title{Extract station matrix from spatialite database}
\usage{
bike_stations(spdb)
}
\arguments{
\item{spdb}{Path to the spatialite database}
}
\value{
Matrix containing data for each station
}
\description{
Extract station matrix from spatialite database
}
