% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-stats.R
\name{bike_summary_stats}
\alias{bike_summary_stats}
\title{Extract summary statistics of database}
\usage{
bike_summary_stats(bikedb)
}
\arguments{
\item{bikedb}{Path to the SQLite3 database}
}
\value{
A \code{data.frame} containing numbers of trips and stations along
with times and dates of first and last trips for each city in database
}
\description{
Extract summary statistics of database
}
