% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-stats.R
\name{bike_total_trips}
\alias{bike_total_trips}
\title{Count number of trips in sqlite3 database}
\usage{
bike_total_trips(bikedb)
}
\arguments{
\item{bikedb}{A string containing the path to the SQLite3 database to 
use.}
}
\description{
Count number of trips in sqlite3 database
}
