% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-test-data.R
\name{bike_write_test_data}
\alias{bike_write_test_data}
\title{Writes test data bundled with package to zip files}
\usage{
bike_write_test_data(data_dir = tempdir())
}
\arguments{
\item{data_dir}{Directory in which data are to be extracted. Defaults to
\code{tempdir()}. If any other directory is specified, files ought to be
removed with \code{bike_rm_test_data()}.}
}
\description{
Writes test data bundled with package to zip files
}
\note{
The entire package works by reading zip-compressed data files provided
by the various hire bicycle systems. Sample data are included with the
package, but these need to be written to zip-compressed archives in order for
the package to be able to use them. This function enables the package to be
tested, both for general use, and for the \code{testhtat} routines.
}
\examples{
\dontrun{
bike_write_test_data ()
list.files (tempdir ())
bike_rm_test_data ()
}
}
