% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikedata-package (Case Conflict).R, R/bikedata-package.R
\docType{package}
\name{bikedata}
\alias{bikedata}
\alias{bikedata-package}
\title{bikedata.}
\description{
bikedata.

bikedata.
}
\author{
Mark Padgham

Mark Padgham
}

