% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_bikedata.R
\name{citibike_files}
\alias{citibike_files}
\title{List of NYC citibike data files}
\usage{
citibike_files()
}
\description{
Ideally the list of file would be automatically extracted, but the files are
stored in an Amazon S3 bucket while requires special tools to read, all of
which require an authentification key to be set up. Present work-around is to
list them here manually, meaning they'll have to be updated on a monthly
basis.
}
