% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createDBIndexes}
\alias{createDBIndexes}
\title{Create indexes in database}
\usage{
createDBIndexes(spdb, tables, cols)
}
\arguments{
\item{spdb}{A string containing the path to the spatialite database to 
use.}

\item{tables}{A vector with the tables for which to create indexes. This
vector should be the same length as the cols vector.}

\item{cols}{A vector with the fields for which to create indexes.}
}
\value{
integer result code
}
\description{
Creates the specified indexes in the database to speed up queries. Note
that for the full dataset this may take some time.
}

