% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripmat.R
\name{filter_tripmat_by_datetime}
\alias{filter_tripmat_by_datetime}
\title{Filter a trip matrix by date and/or time}
\usage{
filter_tripmat_by_datetime(db, ...)
}
\arguments{
\item{db}{The spatiallite database containing the trips (and established by a
\code{dplyr::src_sqlite} call)}

\item{...}{Additional arguments including start_time, end_time, start_date,
end_date, and weekday}
}
\value{
A modified version of the \code{trips} table from \code{db}, filtered
by the specified times
}
\description{
Filter a trip matrix by date and/or time
}

