% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{importDataToSpatialite}
\alias{importDataToSpatialite}
\title{import data using spatialite}
\usage{
importDataToSpatialite(datafiles, spdb)
}
\value{
integer result code
}
\description{
Extracts bike data for NYC citibike
}

