% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lo_stns}
\alias{lo_stns}
\title{Docking stations for London, U.K.}
\format{A \code{data.frame} of the four columns described above.}
\usage{
lo_stns
}
\description{
A \code{data.frame} of station id values, names, and geographic coordinates
for 786 stations for London, U.K. These stations are generally (and by
default) downloaded automatically to ensure they are always up to date, but
such downloading can be disabled in the \code{store_bikedata()} function by
setting \code{latest_lo_stns = FALSE}.
}
\keyword{datasets}
