% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-stats.R
\name{num_trips_in_db}
\alias{num_trips_in_db}
\title{Count number of trips in sqlite3 database}
\usage{
num_trips_in_db(bikedb)
}
\arguments{
\item{bikedb}{A string containing the path to the SQLite3 database to 
use.}
}
\description{
Count number of trips in sqlite3 database
}
