% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_bikedata.R
\name{read_bikedata}
\alias{read_bikedata}
\title{Read nyc-citibike data using C++ routines}
\usage{
read_bikedata(data_dir = tempdir())
}
\arguments{
\item{data_dir}{Directory to which to download the files}
}
\description{
Read nyc-citibike data using C++ routines
}

