% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_bikedata.R
\name{store_bikedata}
\alias{store_bikedata}
\title{Estalish postgres database for nyc-citibike data}
\usage{
store_bikedata(data_dir = tempdir())
}
\arguments{
\item{data_dir}{Directory to which to download the files}
}
\description{
Estalish postgres database for nyc-citibike data
}
