% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripmat.R
\name{tripmat}
\alias{tripmat}
\title{Extract station-to-station trip matrix from spatialite database}
\usage{
tripmat(spdb, quiet = FALSE)
}
\arguments{
\item{spdb}{Path to the spatialite database}

\item{quiet}{If FALSE, progress is displayed on screen}
}
\value{
Square matrix of numbers of trips between each station
}
\description{
Extract station-to-station trip matrix from spatialite database
}
