% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-stats.R
\name{bike_daily_trips}
\alias{bike_daily_trips}
\title{Extract daily trip counts for all stations}
\usage{
bike_daily_trips(bikedb, city, station)
}
\arguments{
\item{bikedb}{A string containing the path to the SQLite3 database.
If no directory specified, it is presumed to be in \code{tempdir()}.}

\item{city}{City for which trips are to be counted - mandatory if database
contains data for more than one city}

\item{station}{Optional argument specifying bike station for which trips are
to be counted}
}
\value{
A \code{data.frame} containing daily dates and total numbers of trips
}
\description{
Extract daily trip counts for all stations
}
\examples{
\dontrun{
bike_write_test_data (data_dir = '.')
# dl_bikedata (city = 'la', data_dir = '.') # or download some real data!
store_bikedata (data_dir = '.', bikedb = 'testdb')
bike_daily_trips (bikedb, city = 'ny')
bike_rm_test_data (data_dir = '.')
bike_rm_db ('testdb')
# don't forget to remove real data!
# file.remove (list.files ('.', pattern = '.zip'))
}
}
