% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-stats.R
\name{bike_daily_trips}
\alias{bike_daily_trips}
\title{Extract daily trip counts for all stations}
\usage{
bike_daily_trips(bikedb, city, station)
}
\arguments{
\item{bikedb}{Path to the SQLite3 database}

\item{city}{City for which trips are to be counted - mandatory if database
contains data for more than one city}

\item{station}{Optional argument specifying bike station for which trips are
to be counted}
}
\value{
A \code{data.frame} containing daily dates and total numbers of trips
}
\description{
Extract daily trip counts for all stations
}
