% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-stats.R
\name{bike_db_totals}
\alias{bike_db_totals}
\title{Count number of entries in sqlite3 database tables}
\usage{
bike_db_totals(bikedb, trips = TRUE, city)
}
\arguments{
\item{bikedb}{A string containing the path to the SQLite3 database.}

\item{trips}{If true, numbers of trips are counted; otherwise numbers of
stations}

\item{city}{Optional city for which numbers of trips are to be counted}
}
\description{
Count number of entries in sqlite3 database tables
}
