% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{bike_stations}
\alias{bike_stations}
\title{Extract station matrix from SQLite3 database}
\usage{
bike_stations(bikedb)
}
\arguments{
\item{bikedb}{Path to the SQLite3 database}
}
\value{
Matrix containing data for each station
}
\description{
Extract station matrix from SQLite3 database
}
