% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_bikedata.R
\name{dl_bikedata}
\alias{dl_bikedata}
\title{Download hire bicycle data}
\usage{
dl_bikedata(city = "nyc", data_dir = tempdir())
}
\arguments{
\item{city}{City for which to download bike data}

\item{data_dir}{Directory to which to download the files}
}
\description{
Download hire bicycle data
}
\section{Details}{

This convenience function downloads hire bicycle data from the nominated
city.  It produces zip-compressed data in R's temporary directory.
Possible cities at present are:
\itemize{
 nyc = New York
}

Ensure you have a fast internet connection and at least 100 Mb space
}

