% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripmat.R
\name{get_datelimits}
\alias{get_datelimits}
\title{Extract date-time limits from trip database}
\usage{
get_datelimits(bikedb)
}
\arguments{
\item{bikedb}{Path to the SQLite3 database}
}
\value{
A vector of 2 elements giving the date-time of the first and last
trips
}
\description{
Extract date-time limits from trip database
}
