% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{importDataToSpatialite}
\alias{importDataToSpatialite}
\title{import data using spatialite}
\usage{
importDataToSpatialite(datafiles, spdb, quiet)
}
\arguments{
\item{datafiles}{A character vector containin the paths to the citibike 
.csv files to import.}

\item{spdb}{A string containing the path to the spatialite database to 
use. It will be created automatically.}

\item{quiet}{If FALSE, progress is displayed on screen}
}
\value{
integer result code
}
\description{
Extracts bike data for NYC citibike
}

