% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_bikedata.R
\name{store_bikedata_spl}
\alias{store_bikedata_spl}
\title{Store data in spatialite database}
\usage{
store_bikedata_spl(datafiles, spdb, quiet = FALSE)
}
\arguments{
\item{datafiles}{A character vector containin the paths to the citibike 
.csv files to import.}

\item{spdb}{A string containing the path to the spatialite database to 
use. It will be created automatically.}

\item{quiet}{If FALSE, progress is displayed on screen}
}
\description{
Store data in spatialite database
}
