% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripmat.R
\name{tripmat}
\alias{tripmat}
\title{Extract station-to-station trip matrix or data.frame from SQLite3 database}
\usage{
tripmat(bikedb, start_date, end_date, start_time, end_time, weekday,
  long = FALSE, quiet = FALSE)
}
\arguments{
\item{bikedb}{Path to the SQLite3 database}

\item{start_date}{If given (as year, month, day) , extract only those records
from and including this date}

\item{end_date}{If given (as year, month, day), extract only those records to
and including this date}

\item{start_time}{If given, extract only those records starting from and
including this time of each day}

\item{end_time}{If given, extract only those records ending at and including
this time of each day}

\item{weekday}{If given, extract only those records including the nominated
weekdays. This can be a vector of numeric, starting with Sunday=1, or
unambiguous characters, so "sa" and "tu" for Saturday and Tuesday.}

\item{long}{If FALSE, a square tripmat of (num-stations, num_stations) is
returns; if TRUE, a long-format matrix of (stn-from, stn-to, ntrips) is
returned.}

\item{quiet}{If FALSE, progress is displayed on screen}
}
\value{
If \code{long=FALSE}, a square matrix of numbers of trips between
each station, otherwise a long-form data.frame with three columns of of
(start_station, end_station, num_trips)
}
\description{
Extract station-to-station trip matrix or data.frame from SQLite3 database
}
\note{
Both dates and times may be given either in numeric or character
format, with arbitrary (or no) delimiters between fields. Single numeric
times are interpreted as hours, with 24 interpreted as day's end at 23:59:59.
}
